+++ package/Tests/Editor/Services/Mock/MockAddOperation.cs
@@ -16,24 +16,18 @@ namespace UnityEditor.PackageManager.UI.Tests
             PackageInfo = packageInfo;
         }
 
-        public void AddPackageAsync(PackageInfo packageInfo, Action<PackageInfo> doneCallbackAction = null, 
+        public void AddPackageAsync(PackageInfo packageInfo, Action<PackageInfo> doneCallbackAction = null,
                                     Action<Error> errorCallbackAction = null)
         {
             if (ForceError != null)
             {
                 if (errorCallbackAction != null)
                     errorCallbackAction(ForceError);
-                
+
                 OnOperationError(ForceError);
             }
             else
             {
-                // Note: TestAddOperation should instead have a list of possible packages to add, and fail if requested package is not in that list
-                if (PackageInfo.PackageId != packageInfo.PackageId)
-                    Debug.LogError(string.Format("Add operation is adding a different package: {0} -- {1}",
-                        packageInfo.PackageId,
-                        PackageInfo.PackageId));
-
                 if (doneCallbackAction != null)
                     doneCallbackAction(PackageInfo);
 
@@ -42,7 +36,7 @@ namespace UnityEditor.PackageManager.UI.Tests
 
             OnOperationFinalized();
         }
-        
+
         internal void ResetEvents()
         {
             OnOperationError = delegate { };