+++ package/Tests/Editor/ValidationSuiteTests/ChangeLogValidationTests.cs
@@ -1,0 +1,174 @@
+using System.IO;
+using NUnit.Framework;
+using UnityEditor.PackageManager.ValidationSuite.ValidationTests;
+using UnityEngine;
+using UnityEditor.PackageManager.ValidationSuite;
+
+
+namespace UnityEditor.PackageManager.ValidationSuite.Tests
+{
+    internal class ChangeLogValidationTests
+    {
+        private const string testDirectory = "tempChangeLogValidationTest";
+
+        [SetUp]
+        public void Setup()
+        {
+
+            if (Directory.Exists(testDirectory))
+            {
+                Directory.Delete(testDirectory, true);
+            }
+            
+            Directory.CreateDirectory(testDirectory);
+        }
+
+        [TearDown]
+        public void TearDown()
+        {
+            if (Directory.Exists(testDirectory))
+            {
+                Directory.Delete(testDirectory, true);
+            }
+        }
+
+        private void CreateChangeLog(string content)
+        {
+            var changeLogPath = Path.Combine(testDirectory, Utilities.ChangeLogFilename);
+            File.AppendAllText(changeLogPath, content);
+        }
+
+        private void CreatePackageJsonFile(string version)
+        {
+            var packageJsonPath = Path.Combine(testDirectory, "package.json");
+            File.WriteAllText(packageJsonPath, "{\"version\":\"" + version + "\"}");
+        }
+        
+        [Test]
+        public void When_ChangeLog_IsMissing_Validation_Fails()
+        {
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Failed, changeLogValiation.TestState);
+            Assert.AreEqual(1, changeLogValiation.TestOutput.Count);
+        }
+
+        [Test]
+        public void When_PackageJson_IsMissing_Validation_Fails()
+        {
+            CreatePackageJsonFile("1.0.0");
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Failed, changeLogValiation.TestState);
+            Assert.AreEqual(1, changeLogValiation.TestOutput.Count);
+        }
+        
+        [Test]
+        public void When_Version_IsMissing_In_ChangeLog_Validation_Fails()
+        {
+            CreatePackageJsonFile("1.0.0");
+            CreateChangeLog("## [2.0.0] - 2033-12-31");
+
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Failed, changeLogValiation.TestState);
+            Assert.AreEqual(1, changeLogValiation.TestOutput.Count);            
+        }
+
+        [Test]
+        public void When_Date_IsMissing_In_ChangeLog_Validation_Fails()
+        {
+            CreatePackageJsonFile("1.0.0");
+            CreateChangeLog("## [2.0.0]");
+
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Failed, changeLogValiation.TestState);
+            Assert.AreEqual(1, changeLogValiation.TestOutput.Count);             
+        }
+
+        [Test]
+        public void When_DateFormat_IsWrong_In_ChangeLog_Validation_Fails()
+        {
+            CreatePackageJsonFile("1.0.0");
+            CreateChangeLog("## [2.0.0] - 31-12-2033");
+
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Failed, changeLogValiation.TestState);
+            Assert.AreEqual(1, changeLogValiation.TestOutput.Count);             
+        }
+
+        [Test]
+        public void When_LogFormat_IsWrong_In_ChangeLog_Validation_Fails()
+        {
+            CreatePackageJsonFile("1.0.0");
+            CreateChangeLog("## 31-12-2033 - [2.0.0]");
+
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Failed, changeLogValiation.TestState);
+            Assert.AreEqual(1, changeLogValiation.TestOutput.Count);             
+        }
+
+        [Test]
+        public void When_Version_and_Date_ArePresent_In_ChangeLog_Validation_Succeeds()
+        {
+            CreatePackageJsonFile("1.0.1");
+            CreateChangeLog("## [1.0.1] - 2033-12-31");
+
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Succeeded, changeLogValiation.TestState);
+            Assert.AreEqual(0, changeLogValiation.TestOutput.Count);            
+        }
+
+        [Test]
+        public void When_Version_and_Date_ArePresent_But_NotFirst_In_ChangeLog_Validation_Fails()
+        {
+            CreatePackageJsonFile("1.0.1");
+            CreateChangeLog("## [5.0.1] - 2033-12-31");
+            CreateChangeLog("## [4.0.1] - 2033-12-31");
+            CreateChangeLog("## [3.0.1] - 2033-12-31");
+            CreateChangeLog("## [2.0.1] - 2033-12-31");
+            CreateChangeLog("## [1.0.1] - 2033-12-31");
+
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Failed, changeLogValiation.TestState);
+            Assert.AreEqual(1, changeLogValiation.TestOutput.Count);             
+        }
+
+        [Test]
+        public void When_Version_and_Date_ArePresent_And_First_In_ChangeLog_Validation_Succeeds()
+        {
+            CreatePackageJsonFile("5.0.1");
+            CreateChangeLog("## [5.0.1] - 2033-12-31");
+            CreateChangeLog("## [4.0.1] - 2033-12-31");
+            CreateChangeLog("## [3.0.1] - 2033-12-31");
+            CreateChangeLog("## [2.0.1] - 2033-12-31");
+            CreateChangeLog("## [1.0.1] - 2033-12-31");
+
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Succeeded, changeLogValiation.TestState);
+            Assert.AreEqual(0, changeLogValiation.TestOutput.Count);             
+        }
+
+        [Test]
+        public void When_Version_and_Date_IsMissing_In_Big_ChangeLog_Validation_Fails()
+        {
+            CreatePackageJsonFile("6.0.1");
+            CreateChangeLog("## [5.0.1] - 2033-12-31");
+            CreateChangeLog("## [4.0.1] - 2033-12-31");
+            CreateChangeLog("## [3.0.1] - 2033-12-31");
+            CreateChangeLog("## [2.0.1] - 2033-12-31");
+            CreateChangeLog("## [1.0.1] - 2033-12-31");
+
+            var changeLogValiation = new ChangeLogValidation();
+            changeLogValiation.Run(testDirectory);
+            Assert.AreEqual(TestState.Failed, changeLogValiation.TestState);
+            Assert.AreEqual(1, changeLogValiation.TestOutput.Count);             
+        }
+    }
+}